factored bits into their own files
